✅ move tests in separate file
