+++ beerus_rest_api/src/lib.rs
@@ -24,276 +24,3 @@ pub async fn build_rocket_server(beerus: BeerusLightClient) -> Rocket<Build> {
 pub fn index() -> &'static str {
     "Hakai!"
 }
-
-#[cfg(test)]
-mod test {
-    use std::str::FromStr;
-
-    use super::rocket;
-    use crate::build_rocket_server;
-    use beerus_core::{
-        config::Config,
-        lightclient::{
-            beerus::BeerusLightClient, ethereum::MockEthereumLightClient,
-            starknet::MockStarkNetLightClient,
-        },
-    };
-    use ethers::types::Address;
-    use primitive_types::U256;
-    use rocket::{http::Status, local::asynchronous::Client};
-    use starknet::core::types::FieldElement;
-
-    /// Test the `query_balance` endpoint.
-    /// `/ethereum/balance/<address>`
-    /// Given normal conditions, when query balance, then ok.
-    #[tokio::test]
-    async fn given_normal_conditions_when_query_balance_then_ok() {
-        // Build mocks.
-        let (config, mut ethereum_lightclient, starknet_lightclient) = config_and_mocks();
-
-        // Given
-        // Mock dependencies.
-        ethereum_lightclient
-            .expect_get_balance()
-            .return_once(move |_, _| Ok(123.into()));
-
-        let beerus = BeerusLightClient::new(
-            config,
-            Box::new(ethereum_lightclient),
-            Box::new(starknet_lightclient),
-        );
-
-        // Build the Rocket instance.
-        let client = Client::tracked(build_rocket_server(beerus).await)
-            .await
-            .expect("valid rocket instance");
-
-        // When
-        let response = client
-            .get(uri!(
-                "/ethereum/balance/0xc24215226336d22238a20a72f8e489c005b44c4a"
-            ))
-            .dispatch()
-            .await;
-
-        // Then
-        assert_eq!(response.status(), Status::Ok);
-        assert_eq!(response.into_string().await.unwrap(), "{\"address\":\"0xc24215226336d22238a20a72f8e489c005b44c4a\",\"balance\":\"0.000000000000000123\",\"unit\":\"ETH\"}");
-    }
-
-    /// Test the `query_balance` endpoint.
-    /// `/ethereum/balance/<address>`
-    /// Given Ethereum light client returns error when query balance, then error is propagated.
-    #[tokio::test]
-    async fn given_ethereum_lightclient_returns_error_when_query_balance_then_error_is_propagated()
-    {
-        // Build mocks.
-        let (config, mut ethereum_lightclient, starknet_lightclient) = config_and_mocks();
-
-        // Given
-        // Mock dependencies.
-        ethereum_lightclient
-            .expect_get_balance()
-            .return_once(move |_, _| Err(eyre::eyre!("cannot query balance")));
-
-        let beerus = BeerusLightClient::new(
-            config,
-            Box::new(ethereum_lightclient),
-            Box::new(starknet_lightclient),
-        );
-
-        // Build the Rocket instance.
-        let client = Client::tracked(build_rocket_server(beerus).await)
-            .await
-            .expect("valid rocket instance");
-
-        // When
-        let response = client
-            .get(uri!(
-                "/ethereum/balance/0xc24215226336d22238a20a72f8e489c005b44c4a"
-            ))
-            .dispatch()
-            .await;
-
-        // Then
-        assert_eq!(response.status(), Status::InternalServerError);
-        assert_eq!(
-            response.into_string().await.unwrap(),
-            "{\"error_message\":\"cannot query balance\"}"
-        );
-    }
-
-    /// Test the `query_starknet_state_root` endpoint.
-    /// `/starknet/state/root`
-    /// Given normal conditions, when query starknet state root, then ok.
-    #[tokio::test]
-    async fn given_normal_conditions_when_query_state_root_then_ok() {
-        // Build mocks.
-        let (config, mut ethereum_lightclient, starknet_lightclient) = config_and_mocks();
-
-        // Expected state root.
-        let expected_starknet_state_root =
-            U256::from_str("0x5bb9692622e817c39663e69dce50777daf4c167bdfa95f3e5cef99c6b8a344d")
-                .unwrap();
-        // Convert to bytes because that's what the mock returns.
-        let mut expected_starknet_state_root_bytes: Vec<u8> = vec![0; 32];
-        expected_starknet_state_root.to_big_endian(&mut expected_starknet_state_root_bytes);
-        // Given
-        // Mock dependencies.
-        ethereum_lightclient
-            .expect_call()
-            .return_once(move |_, _| Ok(expected_starknet_state_root_bytes));
-
-        let beerus = BeerusLightClient::new(
-            config,
-            Box::new(ethereum_lightclient),
-            Box::new(starknet_lightclient),
-        );
-
-        // Build the Rocket instance.
-        let client = Client::tracked(build_rocket_server(beerus).await)
-            .await
-            .expect("valid rocket instance");
-
-        // When
-        let response = client.get(uri!("/starknet/state/root")).dispatch().await;
-
-        // Then
-        assert_eq!(response.status(), Status::Ok);
-        assert_eq!(response.into_string().await.unwrap(), "{\"state_root\":\"2593003852473857760763774375943570015682902311385614557145528717605591462989\"}");
-    }
-
-    /// Test the `query_starknet_state_root` endpoint.
-    /// `/starknet/state/root`
-    /// Given Ethereum light client returns error when query balance, then error is propagated.
-    #[tokio::test]
-    async fn given_ethereum_lightclient_returns_error_when_query_state_root_then_error_is_propagated(
-    ) {
-        // Build mocks.
-        let (config, mut ethereum_lightclient, starknet_lightclient) = config_and_mocks();
-
-        // Given
-        // Mock dependencies.
-        ethereum_lightclient
-            .expect_call()
-            .return_once(move |_, _| Err(eyre::eyre!("cannot query state root")));
-
-        let beerus = BeerusLightClient::new(
-            config,
-            Box::new(ethereum_lightclient),
-            Box::new(starknet_lightclient),
-        );
-
-        // Build the Rocket instance.
-        let client = Client::tracked(build_rocket_server(beerus).await)
-            .await
-            .expect("valid rocket instance");
-
-        // When
-        let response = client.get(uri!("/starknet/state/root")).dispatch().await;
-
-        // Then
-        assert_eq!(response.status(), Status::InternalServerError);
-        assert_eq!(
-            response.into_string().await.unwrap(),
-            "{\"error_message\":\"cannot query state root\"}"
-        );
-    }
-
-    /// Test the `query_starknet_get_storage_at` endpoint.
-    /// `/starknet/storage/<contract>/<key>`
-    /// Given normal conditions, when query starknet get storage at, then ok.
-    #[tokio::test]
-    async fn given_normal_conditions_when_query_starknet_get_storage_at_then_ok() {
-        // Build mocks.
-        let (config, ethereum_lightclient, mut starknet_lightclient) = config_and_mocks();
-
-        // Given
-        let expected_result = FieldElement::from_hex_be("298305742194").unwrap();
-        // Set the expected return value for the StarkNet light client mock.
-        starknet_lightclient
-            .expect_get_storage_at()
-            .times(1)
-            .return_once(move |_address, _key| Ok(expected_result));
-
-        let beerus = BeerusLightClient::new(
-            config,
-            Box::new(ethereum_lightclient),
-            Box::new(starknet_lightclient),
-        );
-
-        // Build the Rocket instance.
-        let client = Client::tracked(build_rocket_server(beerus).await)
-            .await
-            .expect("valid rocket instance");
-
-        // When
-        let response = client.get(uri!("/starknet/storage/0x049d36570d4e46f48e99674bd3fcc84644ddd6b96f7c741b1562b82f9e004dc7/0x341c1bdfd89f69748aa00b5742b03adbffd79b8e80cab5c50d91cd8c2a79be1")).dispatch().await;
-
-        // Then
-        assert_eq!(response.status(), Status::Ok);
-        assert_eq!(
-            response.into_string().await.unwrap(),
-            "{\"result\":\"45642708951444\"}"
-        );
-    }
-
-    /// Test the `query_starknet_get_storage_at` endpoint.
-    /// `/starknet/storage/<contract>/<key>`
-    /// Given StarkNet light client returns error when query starknet get storage at, then error is propagated.
-    #[tokio::test]
-    async fn given_starknet_ligthclient_returns_error_when_query_starknet_get_storage_at_then_error_is_propagated(
-    ) {
-        // Build mocks.
-        let (config, ethereum_lightclient, mut starknet_lightclient) = config_and_mocks();
-
-        // Given
-
-        // Set the expected return value for the StarkNet light client mock.
-        starknet_lightclient
-            .expect_get_storage_at()
-            .times(1)
-            .return_once(move |_address, _key| {
-                Err(eyre::eyre!("cannot query starknet get storage at"))
-            });
-
-        let beerus = BeerusLightClient::new(
-            config,
-            Box::new(ethereum_lightclient),
-            Box::new(starknet_lightclient),
-        );
-
-        // Build the Rocket instance.
-        let client = Client::tracked(build_rocket_server(beerus).await)
-            .await
-            .expect("valid rocket instance");
-
-        // When
-        let response = client.get(uri!("/starknet/storage/0x049d36570d4e46f48e99674bd3fcc84644ddd6b96f7c741b1562b82f9e004dc7/0x341c1bdfd89f69748aa00b5742b03adbffd79b8e80cab5c50d91cd8c2a79be1")).dispatch().await;
-
-        // Then
-        assert_eq!(response.status(), Status::InternalServerError);
-        assert_eq!(
-            response.into_string().await.unwrap(),
-            "{\"error_message\":\"cannot query starknet get storage at\"}"
-        );
-    }
-
-    fn config_and_mocks() -> (Config, MockEthereumLightClient, MockStarkNetLightClient) {
-        let config = Config {
-            ethereum_network: "mainnet".to_string(),
-            ethereum_consensus_rpc: "http://localhost:8545".to_string(),
-            ethereum_execution_rpc: "http://localhost:8545".to_string(),
-            starknet_rpc: "http://localhost:8545".to_string(),
-            starknet_core_contract_address: Address::from_str(
-                "0x0000000000000000000000000000000000000000",
-            )
-            .unwrap(),
-        };
-        (
-            config,
-            MockEthereumLightClient::new(),
-            MockStarkNetLightClient::new(),
-        )
-    }
-}
